fix(groww): widen security_id u32->u64 so Groww 64-bit tokens form candles
Root cause: the in-memory security_id guard was u32, truncating Groww's
64-bit instrument tokens before candle aggregation. Widen ParsedTick /
BufferedSeal / SerializedSeal security_id to u64 across the hot path and
storage; full id round-trips through the 120-128 spill region. Zero DB
migration (LONG column already i64). Verified: trading 1750 + app-groww 72
tests incl bit-62 regression proof; storage 716 lib tests green.

+++ crates/common/src/types.rs
@@ -312,8 +312,14 @@ impl From<&ArchivedOptionType> for OptionType {
     }
 }
 
-/// Unique identifier for a security in the Dhan system.
-pub type SecurityId = u32;
+/// Unique identifier for a security.
+///
+/// Widened u32 → u64 (2026-06-29) so Groww's native exchange_token ids — which
+/// set bit 62 for indices and therefore exceed `u32` — fold through the SAME
+/// pipeline as Dhan. Dhan's wire SecurityId is a 4-byte LE field
+/// (`live-market-feed.md` rule 14) and is widened LOSSLESSLY via `u64::from`
+/// at the parser struct boundary; the wire read width is unchanged.
+pub type SecurityId = u64;
 
 // ---------------------------------------------------------------------------
 // Tests
