+++ crates/storage/src/seal_writer_loop.rs
@@ -176,7 +176,7 @@ mod tests {
         let _ = std::fs::remove_dir_all(dlq);
     }
 
-    fn mk_seal(sid: u32, seg: u8, tf: TfIndex, bucket: u32, close: f64) -> BufferedSeal {
+    fn mk_seal(sid: u64, seg: u8, tf: TfIndex, bucket: u32, close: f64) -> BufferedSeal {
         let mut state = LiveCandleState::empty();
         state.bucket_start_ist_secs = bucket;
         state.open = 100.0;
